--- en.go
@@ -1,36 +1,0 @@
-package timeago
-
-func getEnglish() map[string]string {
-	return map[string]string{
-		"ago":    "ago",
-		"online": "Online",
-		// Seconds
-		"second":   "second",
-		"seconds":  "seconds",
-		"seconds2": "seconds",
-		// Minutes
-		"minute":   "minute",
-		"minutes":  "minutes",
-		"minutes2": "minutes",
-		// Hours
-		"hour":   "hour",
-		"hours":  "hours",
-		"hours2": "hours",
-		// Days
-		"day":   "day",
-		"days":  "days",
-		"days2": "days",
-		// Weeks
-		"week":   "week",
-		"weeks":  "weeks",
-		"weeks2": "weeks",
-		// Months
-		"month":   "month",
-		"months":  "months",
-		"months2": "months",
-		// Years
-		"year":   "year",
-		"years":  "years",
-		"years2": "years",
-	}
-}
