+++ langs/en_test.go
@@ -1,13 +1,13 @@
-package timeago
+package langs
 
 import "testing"
 
 func TestGetEnglish(t *testing.T) {
-	if getEnglish()["weeks2"] != "weeks" {
+	if en["weeks2"] != "weeks" {
 		t.Error("getEnglish must return map of strings with translations for english language")
 	}
 
-	if getEnglish()["days"] != "days" {
+	if en["days"] != "days" {
 		t.Error("getEnglish must return map of strings with translations for english language")
 	}
 }
