+++ langs/lang.go
@@ -1,0 +1,5 @@
+package langs
+
+func GetLanguage(lang string, key string) string {
+	return ""
+}
