+++ langs/ru.go
@@ -1,0 +1,34 @@
+package langs
+
+var ru = map[string]string{
+	"ago":    "назад",
+	"online": "В сети",
+	// Seconds
+	"second":   "секунда",
+	"seconds":  "секунды",
+	"seconds2": "секунд",
+	// Minutes
+	"minute":   "минута",
+	"minutes":  "минуты",
+	"minutes2": "минут",
+	// Hours
+	"hour":   "час",
+	"hours":  "часа",
+	"hours2": "часов",
+	// Days
+	"day":   "день",
+	"days":  "дня",
+	"days2": "дней",
+	// Weeks
+	"week":   "неделя",
+	"weeks":  "недели",
+	"weeks2": "недель",
+	// Months
+	"month":   "месяц",
+	"months":  "месяца",
+	"months2": "месяцев",
+	// Years
+	"year":   "год",
+	"years":  "года",
+	"years2": "лет",
+}
