+++ langs/ru_test.go
@@ -1,13 +1,13 @@
-package timeago
+package langs
 
 import "testing"
 
 func TestGetRussian(t *testing.T) {
-	if getRussian()["weeks2"] != "недель" {
+	if ru["weeks2"] != "недель" {
 		t.Error("getRussian must return map of strings with translations for russian language")
 	}
 
-	if getRussian()["days"] != "дня" {
+	if ru["days"] != "дня" {
 		t.Error("getRussian must return map of strings with translations for russian language")
 	}
 }
